Fix sriov test for mellanox cards
In mellanox when running `ip l show <pf>` the vfs show mac zero.
with this change the test is going to validate that all the
virtual functions are in place before we create a pod.
Then we check the only one virtual function is missing from the host
and return that virtual function number.

Signed-off-by: Sebastian Sch <sebassch@gmail.com>